Tests for frame
